requirements.txt
django
gunicorn
django-heroku
discord.py
requests
psycopg2-binary
dj-database-url
dateparser